Cleanup python logging
In particular, use the logging library's native string interpolation
because it is more performant... it is only evaluated if the log line is
hit, versus building the string ourselves means the string is always
built, regardless of the log level.

Additionally, switched some `log.fatal()` to `log.exception()` so that
it picks up the exception details. This lowers the logging level by one
step, since `log.exception()` is level `ERROR` vs the rarely-used
`log.fatal()` is equal to level `CRITICAL`. But looking at the command
line args, the `quiet` option sets the logger to `WARNING` so both
`ERROR` / `CRITICAL` are always still hit.