Limit the number of files deleted in a single `tarsnap` call to 500
I had a misconfiguration preventing `tarsnapper` from deleting
files several years. When I realized it and fixed it, I discovered I
now have 200K+ tarsnap archives to delete. However, when `tarsnapper`
tries to call `tarsnap` with 200K+ arguments, it results in
`OSError: [Errno 7] Argument list too long`. The root cause is bash
complaining that the bytes length of the entire command exceeds
`ARG_MAX`.

So this limits the maximum number of files deleted in a single call to
`tarsnap` to 500.

500 is a somewhat arbitrary batch size. The actual restriction is
typically a bytes limit on the size of the command that the shell will
accept, which can vary widely across OS flavors. Because it's a bytes
limit, this is also dependent on the length of the backup filenames. So
rather than deal with all this complexity, 500 was chosen as a
reasonable balance between safety and speed.