Ignore codec errors from tarsnap output (#50)
When archiving file names that does not conform to UTF-8, Python would
throw a UnicodeDecodeError, aborting the backup.

Since we are only interested in matching either a text appearing before
tarsnap starting to list files (key file password prompt) or the
finalization of the archival process (EOF), just ignore all conversion
errors to ensure that the archive is properly created.