Exit with non-zero if any job failed or was skipped
Currently any errors coming the underlying tarsnap commands are logged. This is
not threaded back to the entrypoint of the application though, so we end up
exiting with a zero code despite one or more of the jobs having failed. This
makes monitoring the status of tarsnapper impossible. One would have to know to
regularly check its logs to ensure it's actually backing things up.

Fixes #39